# api/docs/swagger.yaml
definitions:
...
  models.CreateStaff:
    properties:
      balance:
        type: number
      branch_id:
        type: string
      name:
        type: string
      password:
        type: string
      tarif_id:
        type: string
      type:
        type: integer
      username:
        type: string
    type: object
  models.CreateStaffTariff:
    properties:
      amountForCard:
        type: number
      amountForCash:
        type: number
      foundedAt:
        type: string
      name:
        type: string
      type:
        description: fixed, percent
        type: integer
    type: object
...
  models.GetAllStaff:
    properties:
      count:
        type: integer
      staffs:
        items:
          $ref: '#/definitions/models.Staff'
        type: array
    type: object
  models.GetAllStaffTariff:
    properties:
      count:
        type: integer
      staffTariffs:
        items:
          $ref: '#/definitions/models.StaffTariff'
        type: array
    type: object
...
  models.Staff:
    properties:
      balance:
        type: number
      branch_id:
        type: string
      created_at:
        type: string
      id:
        type: string
      name:
        type: string
      password:
        type: string
      tarif_id:
        type: string
      type:
        type: integer
      updated_at:
        type: string
      username:
        type: string
    type: object
  models.StaffTariff:
    properties:
      amountForCard:
        type: number
      amountForCash:
        type: number
      createdAt:
        type: string
      foundedAt:
        type: string
      id:
        type: string
      name:
        type: string
      type:
        description: fixed, percent
        type: integer
    type: object
...
  response.ErrorResp:
    properties:
      code:
        type: string
      message:
        type: string
    type: object
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
paths:
...
  /staff:
    get:
      consumes:
      - application/json
      description: get staffs
      parameters:
      - default: 10
        description: limit for response
        in: query
        name: limit
        required: true
        type: integer
      - default: 1
        description: page of req
        in: query
        name: page
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.GetAllStaff'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.ErrorResp'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/response.ErrorResp'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.ErrorResp'
      summary: List staffs
      tags:
      - staffs
    post:
      consumes:
      - application/json
      description: api for create staff
      parameters:
      - description: date of staff
        in: body
        name: staff
        required: true
        schema:
          $ref: '#/definitions/models.CreateStaff'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: string
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.ErrorResp'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/response.ErrorResp'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.ErrorResp'
      summary: create staff
      tags:
      - staffs
  /staff/{id}:
    delete:
      consumes:
      - application/json
      description: api fot delete staff
      parameters:
      - description: id of staff
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: strig
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.ErrorResp'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/response.ErrorResp'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.ErrorResp'
      summary: delete staff
      tags:
      - staffs
    get:
      consumes:
      - application/json
      description: get staff
      parameters:
      - description: id of staff
        format: uuid
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Staff'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.ErrorResp'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/response.ErrorResp'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.ErrorResp'
      summary: get staff
      tags:
      - staffs
    put:
      consumes:
      - application/json
      description: api fot update staffs
      parameters:
      - description: id of staff
        in: path
        name: id
        required: true
        type: string
      - description: id of staff
        in: body
        name: staff
        required: true
        schema:
          $ref: '#/definitions/models.CreateStaff'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: strig
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.ErrorResp'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/response.ErrorResp'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.ErrorResp'
      summary: updateda staff
      tags:
      - staffs
  /staff/Tarif{id}:
    put:
      consumes:
      - application/json
      description: api fot update staffTarif
      parameters:
      - description: id of staffTarif
        in: path
        name: id
        required: true
        type: string
      - description: id of staffTarif
        in: body
        name: staffTarif
        required: true
        schema:
          $ref: '#/definitions/models.CreateStaffTariff'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: strig
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.ErrorResp'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/response.ErrorResp'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.ErrorResp'
      summary: updateda staffTarif
      tags:
      - staffTarifs
  /staffTariff:
    get:
      consumes:
      - application/json
      description: get staffTariff
      parameters:
      - default: 10
        description: limit for response
        in: query
        name: limit
        required: true
        type: integer
      - default: 1
        description: page of req
        in: query
        name: page
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.GetAllStaffTariff'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.ErrorResp'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/response.ErrorResp'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.ErrorResp'
      summary: List staffTariff
      tags:
      - staffTariffs
    post:
      consumes:
      - application/json
      description: api for create staffTariff
      parameters:
      - description: date of staffTariff
        in: body
        name: staffTariff
        required: true
        schema:
          $ref: '#/definitions/models.CreateStaffTariff'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: string
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.ErrorResp'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/response.ErrorResp'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.ErrorResp'
      summary: create staffTariff
      tags:
      - staffTariffs
  /staffTariff/{id}:
    delete:
      consumes:
      - application/json
      description: api fot delete staffTariff
      parameters:
      - description: id of staffTariff
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: strig
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.ErrorResp'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/response.ErrorResp'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.ErrorResp'
      summary: delete staffTariff
      tags:
      - staffTariffs
    get:
      consumes:
      - application/json
      description: get staffTariff
      parameters:
      - description: id of staffTariff
        format: uuid
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.StaffTariff'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.ErrorResp'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/response.ErrorResp'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.ErrorResp'
      summary: get staffTariff
      tags:
      - staffTariffs
...
swagger: "2.0"